apps/data-sync/pyproject.toml
[project]
name = "data-sync"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "azure-functions>=1.17.0",
    "fec-api-client",
    "services",
]

[tool.uv.sources]
fec-api-client = { workspace = true }
services = { workspace = true }
